requirements.txt
altair==4.2.0
anyio==3.6.1
attrs==22.1.0
blinker==1.5
cachetools==5.2.0
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
decorator==5.1.1
entrypoints==0.4
fastapi==0.85.0
gitdb==4.0.9
GitPython==3.1.27
h11==0.13.0
idna==3.4
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.16.0
MarkupSafe==2.1.1
numpy==1.23.3
packaging==21.3
pandas==1.4.4
Pillow==9.2.0
protobuf==3.20.2
pyarrow==9.0.0
pydantic==1.10.2
pydeck==0.8.0b3
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
rich==12.5.1
semver==2.13.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
starlette==0.20.4
toml==0.10.2
toolz==0.12.0
torch==1.12.1+cu113
torchvision==0.13.1+cu113
tornado==6.2
typing_extensions==4.3.0
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.12
uvicorn==0.18.3
validators==0.20.0
watchdog==2.1.9
Werkzeug==2.2.2
zipp==3.8.1
